2023: Refactor header and simplify vars
